split minitwit.java into separate files
