+++ java-itu-minitwit/src/main/java/Utilities/Statics.java
@@ -1,0 +1,8 @@
+package Utilities;
+
+import io.prometheus.client.CollectorRegistry;
+
+public class Statics {
+    public static int latest = 147371;
+
+}
